fixed App for frontend should have new pages in
